+++ attributes/default.rb
@@ -21,6 +21,8 @@ default['oneview']['ruby_sdk_version'] = '~> 4.0'
 default['oneview']['save_resource_info'] = ['uri']
 
 # When looking for a matching Chef resource provider class, this version will be the default.
+# This default value will be overriden if the the property `api_version` is defined or if the client parameter `api_version`
+#   is specified (If the client is a OneviewSDK object, by default it has this parameter defined)
 # A resource provider must be defined for this version. For example, when set to 200, it will look
 # for the resource in OneviewCookbook::API200. When 300, it will look in OneviewCookbook::API300
 # See the libraries/resources directory for more info on supported API versions